Update Main.scala :"run"
addded case class and values:

case class MalformedFile
implicit val decoder
implicit val encoder

"RUN" method read a JSON file containing a Sudoku problem, validate the file and its contents, and attempt to find a solution for the Sudoku problem. If a solution is found, it is printed to the console. If any errors occur during the process, appropriate error messages are raised.